requirements.txt
streamlit
easyocr
pillow
pandas
plost
sklearn
joblib